Made a mistake with github
I committed a .mov file that was far too big along with some other
changes. Github kept failing to sync due to the size of the file so I
tried to revert back a commit and compress the file myself. Instead I
reverted back further than expected because the last 3 or 4 commits
failed to sync.  I deleted the local file and cloned it again.