replaced header with png
I replaced the original header with a png i created in illustrator
because i kept getting an error with the 'billabong' font i imported. I
used the same font in illustrator for continuity with the font in the
instructions.  i also positioned and scaled the instructions to fit over
the window exactly and chose to keep the png transparent so you could
still see the live capture behind the instructions